moved confirmation code
