+++ app/src/main/java/com/finalproject/reminderapp/ui/fragments/HomeFragment.kt
@@ -5,10 +5,13 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.NavHostFragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.finalproject.reminderapp.data.model.Remind
 import com.finalproject.reminderapp.databinding.FragmentHomeBinding
 import com.finalproject.reminderapp.ui.adapter.RemindAdapter
 import com.finalproject.reminderapp.ui.viewModels.HomeViewModel
@@ -56,11 +59,30 @@ class HomeFragment : Fragment() {
     }
 
 
+    fun confirmAndDeleteReminder( remind: Remind) {
+        val builder = AlertDialog.Builder(requireContext())
+        builder.setTitle("Confirm Delete")
+        builder.setMessage("Are you sure you want to delete this reminder?")
+
+        builder.setPositiveButton("Yes") { dialog, _ ->
+            viewModel.deleteReminder(remind)
+            Toast.makeText(requireContext(), "The reminder has been deleted successfully!", Toast.LENGTH_LONG).show()
+            dialog.dismiss()
+        }
+
+        builder.setNegativeButton("No") { dialog, _ ->
+            dialog.dismiss()
+        }
+
+        val deleteConfirmationDialog = builder.create()
+        deleteConfirmationDialog.show()
+    }
+
     fun setupAdapter() {
         adapter = RemindAdapter(
             emptyList(),
             {
-                viewModel.confirmAndDeleteReminder(requireContext(), it)
+                confirmAndDeleteReminder(it)
             },
             {
                 val action = HomeFragmentDirections.actionHomeFragmentToUpdateRemindFragment(it.id!!)
